Refactor Plan to use InterceptorOriginObject instead of Config
Separate out MockLaunchAnglePlanner used in tests
